Custom settings sections. User cluster as plugin
- Plugins can add custom sections to JuliaBox settings screen
- Migrated current implementation of ec2 user clusters as a plugin
