+++ host/tornado/src/juliabox/handlers/__init__.py
@@ -1,5 +1,5 @@
 __author__ = 'tan'
-from handler_base import JBoxHandler, JBoxHandlerPlugin
+from handler_base import JBoxHandler, JBoxHandlerPlugin, JBoxUIModulePlugin
 from admin import AdminHandler
 from auth import AuthHandler
 from cors import CorsHandler
