--- host/tornado/src/juliabox/parallel/__init__.py
@@ -1,2 +1,0 @@
-__author__ = 'tan'
-from user_cluster import UserCluster
