+++ host/tornado/src/juliabox/plugins/parallel/__init__.py
@@ -1,0 +1,3 @@
+__author__ = 'tan'
+from parallel_handler import ParallelHandler, ParallelUIModule
+from parallel_housekeep import ParallelHousekeep
